Remove roles  for mvp

+++ internal/core/modules/member/delete.go
@@ -2,20 +2,31 @@ package member
 
 import (
 	"context"
+	"errors"
 	"fmt"
 
 	"github.com/google/uuid"
-	"github.com/netbill/organizations-svc/internal/core/domain"
 	"github.com/netbill/organizations-svc/internal/core/errx"
-	"github.com/netbill/orgperm"
+	"github.com/netbill/organizations-svc/internal/core/models"
 )
 
 func (m *Module) Delete(
 	ctx context.Context,
-	initiator domain.AccountActor,
+	actor models.AccountActor,
 	memberID uuid.UUID,
 ) error {
 	member, err := m.GetByID(ctx, memberID)
+	if errors.Is(err, errx.ErrorMemberNotExists) {
+		buried, err := m.repo.MemberIsBuried(ctx, memberID)
+		if err != nil {
+			return err
+		}
+		if buried {
+			return errx.ErrorMemberDeleted.Raise(
+				fmt.Errorf("member with id %s is already deleted", memberID),
+			)
+		}
+	}
 	if err != nil {
 		return err
 	}
@@ -25,18 +36,44 @@ func (m *Module) Delete(
 		)
 	}
 
-	err = m.checkAbilityToDeleteMember(ctx, initiator, member.OrganizationID, memberID)
+	if member.AccountID == actor {
+		return errx.ErrorCannotDeleteSelf.Raise(
+			fmt.Errorf("account cannot delete itself as member %s", member.ID),
+		)
+	}
+
+	organization, err := m.repo.GetOrganizationByID(ctx, member.OrganizationID)
 	if err != nil {
 		return err
 	}
+	if organization.Status == models.OrganizationStatusSuspended {
+		return errx.ErrorOrganizationIsSuspended.Raise(
+			fmt.Errorf("organization %s is suspended", member.OrganizationID),
+		)
+	}
+
+	initiator, err := m.getInitiator(ctx, actor, member.OrganizationID)
+	if err != nil {
+		return err
+	}
+	if !initiator.Head {
+		return errx.ErrorNotOrganizationHead.Raise(
+			fmt.Errorf("account has no rights to delete member %s", memberID),
+		)
+	}
 
 	return m.repo.Transaction(ctx, func(ctx context.Context) error {
+		err = m.repo.BuryMember(ctx, memberID)
+		if err != nil {
+			return fmt.Errorf("failed to bury member %s: %w", memberID, err)
+		}
+
 		err = m.repo.DeleteMember(ctx, memberID)
 		if err != nil {
 			return fmt.Errorf("failed to delete member %s: %w", memberID, err)
 		}
 
-		if err = m.messenger.WriteOrgMemberDeleted(ctx, member.ID); err != nil {
+		if err = m.messenger.WriteOrgMemberDeleted(ctx, memberID); err != nil {
 			return fmt.Errorf("failed to send member deleted message for member %s: %w", memberID, err)
 		}
 
@@ -44,50 +81,35 @@ func (m *Module) Delete(
 	})
 }
 
-func (m *Module) checkAbilityToDeleteMember(
-	ctx context.Context,
-	initiator domain.AccountActor,
-	organizationID uuid.UUID,
-	memberID uuid.UUID,
-) error {
-	member, err := m.GetInitiator(ctx, initiator, organizationID)
+func (m *Module) DeleteSelf(ctx context.Context, actor models.AccountActor, orgID uuid.UUID) error {
+	_, err := m.repo.GetOrganizationByID(ctx, orgID)
 	if err != nil {
 		return err
 	}
 
-	if !member.Head {
-		hasPermission, err := m.repo.CheckMemberHavePermission(ctx, member.AccountID, orgperm.MembersDeleteID)
-		if err != nil {
-			return err
-		}
-		if !hasPermission {
-			return errx.ErrorNotEnoughRights.Raise(
-				fmt.Errorf("initiator member %s has no delete members permission", member.ID),
-			)
-		}
-
-		firstMaxRole, err := m.repo.GetMemberMaxRole(ctx, member.ID)
-		if err != nil {
-			return fmt.Errorf("failed to get max role for member %s: %w", member.AccountID, err)
-		}
-
-		secMaxRole, err := m.repo.GetMemberMaxRole(ctx, memberID)
-		if err != nil {
-			return fmt.Errorf("failed to get max role for member %s: %w", memberID, err)
-		}
-
-		if firstMaxRole.Rank < secMaxRole.Rank {
-			return errx.ErrorNotEnoughRights.Raise(
-				fmt.Errorf(
-					"member %s with rank %d cannot manage member %s with rank %d",
-					member.AccountID,
-					firstMaxRole.Rank,
-					memberID,
-					secMaxRole.Rank,
-				),
-			)
-		}
+	member, err := m.getInitiator(ctx, actor, orgID)
+	if err != nil {
+		return err
+	}
+	if member.Head {
+		return errx.ErrorCannotDeleteOrganizationHeadMember.Raise(
+			fmt.Errorf("cannot delete organization head member %s", member.ID),
+		)
 	}
 
-	return nil
+	return m.repo.Transaction(ctx, func(ctx context.Context) error {
+		if err = m.repo.BuryMember(ctx, member.ID); err != nil {
+			return fmt.Errorf("failed to bury member %s: %w", member.ID, err)
+		}
+
+		if err = m.repo.DeleteMember(ctx, member.ID); err != nil {
+			return fmt.Errorf("failed to delete member %s: %w", member.ID, err)
+		}
+
+		if err = m.messenger.WriteOrgMemberDeleted(ctx, member.ID); err != nil {
+			return fmt.Errorf("failed to send member deleted message for member %s: %w", member.ID, err)
+		}
+
+		return nil
+	})
 }
