+++ internal/core/modules/organization/upload_media.go
@@ -5,49 +5,108 @@ import (
 	"fmt"
 
 	"github.com/google/uuid"
-	"github.com/netbill/organizations-svc/internal/core/domain"
+	"github.com/netbill/organizations-svc/internal/core/errx"
+	"github.com/netbill/organizations-svc/internal/core/models"
 )
 
 func (m *Module) CreateOrgUploadMediaLinks(
 	ctx context.Context,
-	actor domain.AccountActor,
+	actor models.AccountActor,
 	organizationID uuid.UUID,
-) (domain.Organization, domain.UploadOrgMediaLinks, error) {
+) (models.Organization, models.UploadOrgMediaLinks, error) {
 	org, err := m.GetByID(ctx, organizationID)
 	if err != nil {
-		return domain.Organization{}, domain.UploadOrgMediaLinks{}, err
+		return models.Organization{}, models.UploadOrgMediaLinks{}, err
 	}
 
-	err = m.chekPermissionForManageOrganization(ctx, actor, org.ID)
+	member, err := m.getInitiator(ctx, actor, organizationID)
 	if err != nil {
-		return domain.Organization{}, domain.UploadOrgMediaLinks{}, err
+		return models.Organization{}, models.UploadOrgMediaLinks{}, err
+	}
+	if !member.Head {
+		return models.Organization{}, models.UploadOrgMediaLinks{}, errx.ErrorNotOrganizationHead.Raise(
+			fmt.Errorf("only organization head member can activate organization, but member %s is not head", member.ID),
+		)
+	}
+
+	if org.Status == models.OrganizationStatusSuspended {
+		return models.Organization{}, models.UploadOrgMediaLinks{}, errx.ErrorOrganizationIsSuspended.Raise(
+			fmt.Errorf("organization with id %s is suspended", organizationID),
+		)
 	}
 
 	iconLinks, err := m.bucket.CreateOrganizationIconUploadMediaLinks(ctx, organizationID)
 	if err != nil {
-		return domain.Organization{}, domain.UploadOrgMediaLinks{}, fmt.Errorf("failed to create upload media links: %w", err)
+		return models.Organization{}, models.UploadOrgMediaLinks{}, fmt.Errorf("failed to create upload media links: %w", err)
 	}
 
 	bannerLinks, err := m.bucket.CreateOrganizationBannerUploadMediaLinks(ctx, organizationID)
 	if err != nil {
-		return domain.Organization{}, domain.UploadOrgMediaLinks{}, fmt.Errorf("failed to create upload media links: %w", err)
+		return models.Organization{}, models.UploadOrgMediaLinks{}, fmt.Errorf("failed to create upload media links: %w", err)
 	}
 
-	return org, domain.UploadOrgMediaLinks{
+	return org, models.UploadOrgMediaLinks{
 		Icon:   iconLinks,
 		Banner: bannerLinks,
 	}, nil
 }
 
+func (m *Module) updateOrganizationIcon(
+	ctx context.Context,
+	organization models.Organization,
+	params UpdateParams,
+) (newKey *string, err error) {
+	if params.IconKey != nil {
+		if err = m.bucket.ValidateOrganizationIcon(ctx, organization.ID, *params.IconKey); err != nil {
+			return nil, fmt.Errorf("failed to validate organization icon: %w", err)
+		}
+
+		iconKey, err := m.bucket.UpdateOrganizationIcon(ctx, organization.ID, *params.IconKey)
+		if err != nil {
+			return nil, fmt.Errorf("failed to update organization icon: %w", err)
+		}
+
+		if err = m.bucket.DeleteOrganizationIcon(ctx, organization.ID, iconKey); err != nil {
+			return nil, fmt.Errorf("failed to delete organization icon: %w", err)
+		}
+
+		newKey = &iconKey
+	}
+
+	if organization.IconKey != nil {
+		if err = m.bucket.DeleteOrganizationIcon(ctx, organization.ID, *organization.IconKey); err != nil {
+			return nil, fmt.Errorf("failed to delete organization icon: %w", err)
+		}
+	}
+
+	return newKey, nil
+}
+
 func (m *Module) DeleteOrgUploadIcon(
 	ctx context.Context,
-	actor domain.AccountActor,
+	actor models.AccountActor,
 	organizationID uuid.UUID,
 	key string,
 ) error {
-	err := m.chekPermissionForManageOrganization(ctx, actor, organizationID)
+	org, err := m.GetByID(ctx, organizationID)
 	if err != nil {
 		return err
+	}
+
+	if org.Status == models.OrganizationStatusSuspended {
+		return errx.ErrorOrganizationIsSuspended.Raise(
+			fmt.Errorf("organization with id %s is suspended", organizationID),
+		)
+	}
+
+	member, err := m.getInitiator(ctx, actor, organizationID)
+	if err != nil {
+		return err
+	}
+	if !member.Head {
+		return errx.ErrorNotOrganizationHead.Raise(
+			fmt.Errorf("only organization head member can activate organization, but member %s is not head", member.ID),
+		)
 	}
 
 	err = m.bucket.DeleteUploadOrganizationIcon(ctx, organizationID, key)
@@ -58,15 +117,62 @@ func (m *Module) DeleteOrgUploadIcon(
 	return nil
 }
 
+func (m *Module) updateOrganizationBanner(
+	ctx context.Context,
+	organization models.Organization,
+	params UpdateParams,
+) (newKey *string, err error) {
+	if params.BannerKey != nil {
+		if err = m.bucket.ValidateOrganizationBanner(ctx, organization.ID, *params.BannerKey); err != nil {
+			return nil, fmt.Errorf("failed to validate organization banner: %w", err)
+		}
+
+		key, err := m.bucket.UpdateOrganizationBanner(ctx, organization.ID, *params.BannerKey)
+		if err != nil {
+			return nil, fmt.Errorf("failed to update organization banner: %w", err)
+		}
+
+		if err = m.bucket.DeleteOrganizationBanner(ctx, organization.ID, key); err != nil {
+			return nil, fmt.Errorf("failed to delete organization banner: %w", err)
+		}
+
+		newKey = &key
+	}
+
+	if organization.BannerKey != nil {
+		if err = m.bucket.DeleteOrganizationBanner(ctx, organization.ID, *organization.BannerKey); err != nil {
+			return nil, fmt.Errorf("failed to delete organization banner: %w", err)
+		}
+	}
+
+	return newKey, nil
+}
+
 func (m *Module) DeleteOrgUploadBanner(
 	ctx context.Context,
-	actor domain.AccountActor,
+	actor models.AccountActor,
 	organizationID uuid.UUID,
 	key string,
 ) error {
-	err := m.chekPermissionForManageOrganization(ctx, actor, organizationID)
+	org, err := m.GetByID(ctx, organizationID)
 	if err != nil {
 		return err
+	}
+
+	if org.Status == models.OrganizationStatusSuspended {
+		return errx.ErrorOrganizationIsSuspended.Raise(
+			fmt.Errorf("organization with id %s is suspended", organizationID),
+		)
+	}
+
+	member, err := m.getInitiator(ctx, actor, organizationID)
+	if err != nil {
+		return err
+	}
+	if !member.Head {
+		return errx.ErrorNotOrganizationHead.Raise(
+			fmt.Errorf("only organization head member can activate organization, but member %s is not head", member.ID),
+		)
 	}
 
 	err = m.bucket.DeleteUploadOrganizationBanner(ctx, organizationID, key)
